requirements.txt
numpy==1.23.4
PyYAML==6.0
matplotlib==3.6.2
scikit-image==0.19.3
setuptools==60.2.0
tabulate==0.9.0
scipy==1.9.3
scikit-learn==1.1.3
torch==1.12.0
torchvision==0.13.0
pytorch-lightning==1.8.1
torchinfo==1.7.1

